Run validations concurrently
This change runs the three validations concurrently, which is not the
default when combining IO values applicatively.

It’s a bit frustrating to have to call parMapN and then also mapN: it
feels like it should be possible to avoid the two by using a parallel
applicative. I tried this (using ParallelF.apply and ParallelF.value to
convert from IO to the Par newtype for which the Applicative instance
combines actions concurrently), but it was more onerous than expected
because I had to convert each action on the way in and the result
afterwards as well, along with the verbose applicative composition
syntax:

```scala
ParallelF.value(
  Applicative[IO.Par]
    .compose[[A] =>> ValidatedNec[UserError, A]](
      ParallelF(validateName(name)),
      ParallelF(validateEmail(email)),
      ParallelF(IO.pure(validateNumber(number)))
    )(UserData.apply)
)
```

…so I decided this was better despite my frustration.